// src/dtos/comments/editComment.dto.ts
import z from "zod";

export interface EditCommentInputDTO {
  token: string;
  commentId: string;
  content: string;
}

export type EditCommentOutputDTO = undefined;

export const EditCommentSchema = z
  .object({
    token: z.string().min(1),
    commentId: z.string().min(1),
    content: z.string().min(1),
  })
  .transform((data) => data as EditCommentInputDTO);